requirements.txt
py-slippi==1.6.2